cloudrun/customer_ai/requirements.txt
flask==3.0.3
google-cloud-bigquery==3.25.0
google-cloud-aiplatform==1.60.0
vertexai==1.60.0
gunicorn==22.0.0
